rms native/ module, condensing files and moving into io/
