--- py4DSTEM/io/native/__init__.py
@@ -1,3 +1,0 @@
-from py4DSTEM.io.native.read import *
-from py4DSTEM.io.native.write import *
-
