+++ py4DSTEM/io/read.py
@@ -1,11 +1,13 @@
 # Functions for reading native and non-native file types
 
-import h5py
 import pathlib
-from os.path import exists, splitext
+from os.path import exists, splitext, basename, dirname, join
 from typing import Union, Optional
+import warnings
+import h5py
+import numpy as np
 
-from py4DSTEM.io.native.read import read_py4DSTEM
+# Readers for non-native filetypes
 from py4DSTEM.io.nonnative import (
     read_empad,
     read_dm,
@@ -13,9 +15,34 @@ from py4DSTEM.io.nonnative import (
     load_mib
 )
 
+# Classes for native reader
+from py4DSTEM.io.datastructure import (
+    Root,
+    Tree,
+    ParentTree,
+    Metadata,
+    Array,
+    PointList,
+    PointListArray,
+    Calibration,
+    DataCube,
+    DiffractionSlice,
+    VirtualDiffraction,
+    RealSlice,
+    VirtualImage,
+    Probe,
+    QPoints,
+    BraggVectors
+)
 
 
-# Parse filetypes
+
+
+
+
+
+###### GENERAL UTILITY FUNCTIONS ######
+
 
 def parse_filetype(fp):
     """ Accepts a path to a 4D-STEM dataset, and returns the file type.
@@ -59,8 +86,10 @@ def parse_filetype(fp):
 
 
 
-# Read non-native files by
-# parsing the filetype and calling the appropriate reader
+
+
+###### NON-NATIVE FILE READER ######
+
 
 def import_file(
     filepath: Union[str, pathlib.Path],
@@ -71,6 +100,7 @@ def import_file(
 ):
     """
     Reader for non-native file formats.
+    Parses the filetype, and calls the appropriate reader.
     Supports Gatan DM3/4, some EMPAD file versions, Gatan K2 bin/gtg, and mib
     formats.
 
@@ -83,10 +113,9 @@ def import_file(
             from storage.
         binfactor (int): Diffraction space binning factor for bin-on-load.
         filetype (str): Used to override automatic filetype detection.
-        **kwargs:
-
-    For documentation of kwargs, refer to the individual readers, in
-    py4DSTEM.io.
+        **kwargs: any additional kwargs are passed to the downstream reader -
+            refer to the individual filetype reader function call signatures
+            and docstrings for more details.
 
     Returns:
         (DataCube or Array) returns a DataCube if 4D data is found, otherwise
@@ -140,24 +169,181 @@ def import_file(
 
 
 
-# Read native files by
-# error checking the inputs then passing to another function ... :'(
+
+
+
+
+
+
+###### NATIVE FILE READER #######
+
+
+
+# utility functions
+
+def get_py4DSTEM_topgroups(filepath):
+    """
+    Returns a list of toplevel groups in an HDF5 file which are valid
+    py4DSTEM file trees.
+    """
+    valid_emd_keys = (0,1,2,3,'root')
+
+    topgroups = []
+    with h5py.File(filepath,'r') as f:
+        for key in f.keys():
+            if 'emd_group_type' in f[key].attrs:
+                if f[key].attrs['emd_group_type'] in valid_emd_keys:
+                    topgroups.append(key)
+
+    return topgroups
+
+def is_py4DSTEM_file(filepath):
+    """
+    Returns True iff filepath points to a py4DSTEM formatted (EMD type 2) file.
+    """
+    try:
+        topgroups = get_py4DSTEM_topgroups(filepath)
+        if len(topgroups)>0:
+            return True
+        else:
+            return False
+    except OSError:
+        return False
+
+def get_py4DSTEM_version(filepath, topgroup=None):
+    """
+    Returns the version (major,minor,release) of a py4DSTEM file.
+    """
+    assert(is_py4DSTEM_file(filepath)), "Error: not recognized as a py4DSTEM file"
+    with h5py.File(filepath,'r') as f:
+        if topgroup is None or topgroup not in f.keys():
+            if "4DSTEM" in f.keys():
+                topgroup = "4DSTEM"
+            elif "4DSTEM_experiment" in f.keys():
+                topgroup = "4DSTEM_experiment"
+            else:
+                raise ValueError("no root group exists with default names. please specify topgroup.")
+        version_major = int(f[topgroup].attrs['version_major'])
+        version_minor = int(f[topgroup].attrs['version_minor'])
+        if 'version_release' in f[topgroup].attrs.keys():
+            version_release = int(f[topgroup].attrs['version_release'])
+        else:
+            version_release = 0
+        return version_major, version_minor, version_release
+
+def get_UUID(filepath, topgroup='4DSTEM'):
+    """
+    Returns the UUID of a py4DSTEM file, or if unavailable returns -1.
+    """
+    assert(is_py4DSTEM_file(filepath)), "Error: not recognized as a py4DSTEM file"
+    with h5py.File(filepath,'r') as f:
+        if topgroup in f.keys():
+            if 'UUID' in f[topgroup].attrs:
+                return f[topgroup].attrs['UUID']
+    return -1
+
+def version_is_geq(current,minimum):
+    """
+    Returns True iff current version (major,minor,release) is greater than or equal to minimum."
+    """
+    if current[0]>minimum[0]:
+        return True
+    elif current[0]==minimum[0]:
+        if current[1]>minimum[1]:
+            return True
+        elif current[1]==minimum[1]:
+            if current[2]>=minimum[2]:
+                return True
+        else:
+            return False
+    else:
+        return False
+
+def _get_class(grp):
+
+    lookup = {
+        'Metadata' : Metadata,
+        'Array' : Array,
+        'PointList' : PointList,
+        'PointListArray' : PointListArray,
+        'Calibration' : Calibration,
+        'DataCube' : DataCube,
+        'DiffractionSlice' : DiffractionSlice,
+        'VirtualDiffraction' : VirtualDiffraction,
+        'DiffractionImage' : VirtualDiffraction,
+        'RealSlice' : RealSlice,
+        'VirtualImage' : VirtualImage,
+        'Probe' : Probe,
+        'QPoints' : QPoints,
+        'BraggVectors' : BraggVectors
+    }
+    try:
+        classname = grp.attrs['py4dstem_class']
+        __class__ = lookup[classname]
+        return __class__
+    except KeyError:
+        return None
+        #raise Exception(f"Unknown classname {classname}")
+
+
+# Print the HDF5 filetree to screen
+
+def print_h5_tree(filepath, show_metadata=False):
+    """
+    Prints the contents of an h5 file from a filepath.
+    """
+
+    with h5py.File(filepath,'r') as f:
+        print('/')
+        print_h5pyFile_tree(f, show_metadata=show_metadata)
+        print('\n')
+
+def print_h5pyFile_tree(f, tablevel=0, linelevels=[], show_metadata=False):
+    """
+    Prints the contents of an h5 file from an open h5py File instance.
+    """
+    if tablevel not in linelevels:
+        linelevels.append(tablevel)
+    keys = [k for k in f.keys() if isinstance(f[k],h5py.Group)]
+    if not show_metadata:
+        keys = [k for k in keys if k != '_metadata']
+    N = len(keys)
+    for i,k in enumerate(keys):
+        string = ''
+        string += '|' if 0 in linelevels else ''
+        for idx in range(tablevel):
+            l = '|' if idx+1 in linelevels else ''
+            string += '\t'+l
+        #print(string)
+        print(string+'--'+k)
+        if i == N-1:
+            linelevels.remove(tablevel)
+        print_h5pyFile_tree(
+            f[k],
+            tablevel=tablevel+1,
+            linelevels=linelevels,
+            show_metadata=show_metadata)
+
+    pass
+
+
+
+# native file reader
 
 def read(
-    filepath: Union[str,pathlib.Path],
+    filepath,
     root: Optional[str] = None,
     tree: Optional[Union[bool,str]] = True,
-    **kwargs,
+    **legacy_options,
     ):
     """
-    File reader for files written by py4DSTEM. To load non-native
-    file types, use py4DSTEM.import_file.
+    File reader for files *written by* py4DSTEM.
+    To read non-native file formats, please use py4DSTEM.import_file.
 
     For files written by py4DSTEM v0.13+, the arguments this function
-    accepts and their behaviors are below. For older verions, see
-    the docstring for `py4DSTEM.io.native.legacy.read_py4DSTEM_legacy`
-    for keyword arguments and their behaviors.
-
+    accepts and their behaviors are below. For older verions, pass
+    kwargs (**legacy_options) according to the legacy reader args
+    found in py4DSTEM.io.legacy.read_py4DSTEM_legacy`.
 
     Args:
         filepath (str or Path): the file path
@@ -182,26 +368,184 @@ def read(
             loading the whole datacube).
 
     Returns:
-        (the data)
+        (data, variable type) the form of the returned data depends on
+            the `tree` argument - see it's docs, above.
     """
+    # parse filepath
+    # (must be valid, exist, and point to a py4DSTEM file)
+    er1 = f"filepath must be a string or Path, not {type(filepath)}"
+    assert(isinstance(filepath, (str,pathlib.Path) )), er1
+    er2 = f"specified filepath '{filepath}' does not exist"
+    assert(exists(filepath)), er2
+    er3 = f"Error: {filepath} isn't recognized as a py4DSTEM file."
+    assert(is_py4DSTEM_file(filepath)), er3
 
     # parse filetype
-    er1 = f"filepath must be a string or Path, not {type(filepath)}"
-    er2 = f"specified filepath '{filepath}' does not exist"
-    assert(isinstance(filepath, (str,pathlib.Path) )), er1
-    assert(exists(filepath)), er2
-
     filetype = parse_filetype(filepath)
-    assert filetype == "py4DSTEM", "Incompatible file type for py4DSTEM.io.read. To non-native files must be read with py4DSTEM.io.import_file"
+    assert filetype == "py4DSTEM", "Non-native file type detected. To import data from non-native formats, use p4DSTEM.import_file"
 
-    # prepare kwargs
-    kwargs['root'] = root
-    kwargs['tree'] = tree
 
-    # load data
-    data = read_py4DSTEM(
-        filepath,
-        **kwargs
-    )
+    # Check the EMD version
+    if root is not None:
+        v = get_py4DSTEM_version(filepath, root.split("/")[0])
+    else:
+        try:
+            with h5py.File(filepath,'r') as f:
+                k = list(f.keys())[0]
+                v = get_py4DSTEM_version(filepath, k)
+        except:
+            raise Exception('error parsing file version...')
 
+
+    # Use legacy readers for EMD files with v<13
+    if v[1] <= 12:
+        from py4DSTEM.io.native.legacy import read_py4DSTEM_legacy
+        return read_py4DSTEM_legacy(filepath,**legacy_options)
+
+
+    # Read EMD files with v>=13
+
+    # if root is None, determine if there is a single object in the file
+    # if so, set root to that file; otherwise raise an Exception or Warning
+    if root is None:
+        with h5py.File(filepath,'r') as f:
+            l1keys = list(f.keys())
+            if len(l1keys)==0:
+                raise Exception('No top level groups found in this HDF5 file!')
+            elif len(l1keys)>1:
+                warnings.warn('Multiple top level groups found; please specify. Returning group names.')
+                return l1keys
+            else:
+                l2keys = list(f[l1keys[0]].keys())
+                if len(l2keys)==0:
+                    raise Exception('No top level data blocks found in this HDF5 file!')
+                elif len(l2keys)>1:
+                    warnings.warn('Multiple top level data blocks found; please specify. Returning h5 paths to top level data blocks.')
+                    return [join(l1keys[0],k) for k in l2keys]
+                else:
+                    root = join(l1keys[0],l2keys[0])
+                    #this is a windows fix
+                    root = root.replace("\\","/")
+
+    # Open h5 file
+    with h5py.File(filepath,'r') as f:
+
+        # Open the selected group
+        try:
+            group_data = f[root]
+
+            # Read for...
+
+            # ...the whole file tree
+            if tree is True:
+                return _read_with_tree(group_data)
+
+            # ...the specified node only
+            elif tree is False:
+                return _read_without_tree(group_data)
+
+            # ...the tree under the specified node,
+            # excluding the node itself
+            elif tree == 'noroot':
+                return _read_without_root(group_data)
+
+        except KeyError:
+            raise Exception(f"the provided root {root} is not a valid path to a recognized data group")
+
+
+
+# utilities for performing the read in each use-case,
+# walking or not walking the tree as specified,
+# and attaching calibrations where appropriate
+
+def _read_without_tree(grp):
+
+    # handle empty datasets
+    if grp.attrs['emd_group_type'] == 'root':
+        data = Root(
+            name = basename(grp.name),
+        )
+        cal = _add_calibration(
+            data.tree,
+            grp
+        )
+        if cal is not None:
+            data.tree = ParentTree(data, cal)
+            data.calibration = cal
+        return data
+
+    # read all other data
+    __class__ = _get_class(grp)
+    data = __class__.from_h5(grp)
+    if not isinstance(data, Calibration):
+        cal = _add_calibration(
+            data.tree,
+            grp
+        )
+        if cal is not None:
+            data.tree = ParentTree(data, cal)
+            data.calibration = cal
     return data
+
+
+def _read_with_tree(grp):
+    data = _read_without_tree(grp)
+    _populate_tree(
+        data.tree,
+        grp
+    )
+    return data
+
+
+def _read_without_root(grp):
+    root = Root()
+    cal = _add_calibration(
+        root,
+        grp
+    )
+    if cal is not None:
+        root.tree = ParentTree(root,cal)
+        root.calibration = cal
+    _populate_tree(
+        root.tree,
+        grp
+    )
+    return root
+
+
+# TODO: case of multiple Calibration instances
+def _add_calibration(tree,grp):
+    keys = [k for k in grp.keys() if isinstance(grp[k],h5py.Group)]
+    keys = [k for k in keys if (_get_class(grp[k]) == Calibration)]
+    if len(keys)>0:
+        k = keys[0]
+        tree[k] = _read_without_tree(grp[k])
+        return tree[k]
+    else:
+        name = dirname(grp.name)
+        if name != '/':
+            grp_upstream = grp.file[dirname(grp.name)]
+            return _add_calibration(tree, grp_upstream)
+        else:
+            return None
+
+
+def _populate_tree(tree,grp):
+    keys = [k for k in grp.keys() if isinstance(grp[k],h5py.Group)]
+    keys = [k for k in keys if (k[0] != '_' and not _get_class(
+                grp[k]) == Calibration)]
+    for key in keys:
+        tree[key] = _read_without_tree(
+            grp[key]
+        )
+        _populate_tree(
+            tree[key].tree,
+            grp[key]
+        )
+    pass
+
+
+
+
+
+
